Add a model to a tidyflow

plug_model() adds a parsnip model to the tidyflow.

drop_model() removes the model specification as well as any fitted model object. Any extra formulas are also removed. Doesn't remove any steps from the pre stage.

replace_model() first removes the model then adds the new specification to the tidyflow.

plug_model(x, spec, formula = NULL)

drop_model(x)

replace_model(x, spec, formula = NULL)

Arguments

x
A tidyflow.

spec
A parsnip model specification.

formula
An optional formula override to specify the terms of the model. Typically, the terms are extracted from the formula or recipe preprocessing methods. However, some models (like survival and bayesian models) use the formula not to preprocess, but to specify the structure of the model. In those cases, a formula specifying the model structure must be passed unchanged into the model call itself. This argument is used for those purposes.

Value

x, updated with either a new or removed model.

Details

plug_model() is a required step to construct a minimal tidyflow.

Examples

library(parsnip)

# Define two competing model:
lm_model <- set_engine(linear_reg(), "lm")
regularized_model <- set_engine(lm_model, "glmnet")

# Define a minimal tidyflow: data + formula + model
wf <-
  mtcars %>%
  tidyflow() %>%
  plug_formula(mpg ~ .) %>% 
  plug_model(lm_model)

wf

#> ══ Tidyflow ════════════════════════════════════════════════════════════════════
#> Data: 32 rows x 11 columns
#> Split: None
#> Formula: mpg ~ .
#> Resample: None
#> Grid: None
#> Model:
#> Linear Regression Model Specification (regression)
#> 
#> Computational engine: lm 
#> 

# We can drop the model at any time and the remaining steps
# are intact
drop_model(wf)

#> ══ Tidyflow ════════════════════════════════════════════════════════════════════
#> Data: 32 rows x 11 columns
#> Split: None
#> Formula: mpg ~ .
#> Resample: None
#> Grid: None
#> Model: None

# We can fit the model with `fit`:
fitted <- fit(wf)

# Extract the model if needed:
fitted %>%
  pull_tflow_fit()

#> parsnip model object
#> 
#> Fit time:  6ms 
#> 
#> Call:
#> stats::lm(formula = ..y ~ ., data = data)
#> 
#> Coefficients:
#> (Intercept)          cyl         disp           hp         drat           wt  
#>    12.30337     -0.11144      0.01334     -0.02148      0.78711     -3.71530  
#>        qsec           vs           am         gear         carb  
#>     0.82104      0.31776      2.52023      0.65541     -0.19942  
#> 

# If we remove the model from the fitted `tidyflow`,
# the fit is dropped:
drop_model(fitted)

#> ══ Tidyflow ════════════════════════════════════════════════════════════════════
#> Data: 32 rows x 11 columns
#> Split: None
#> Formula: mpg ~ .
#> Resample: None
#> Grid: None
#> Model: None

# We could replace the model from the initial tidyflow with
# the regularized model with `replace_model`

## TODO: when https://github.com/cimentadaj/tidyflow/issues/4 is fixed
## replace wf with fitted here.

reg_fitted <-
  wf %>%
  replace_model(regularized_model) %>%
  fit()

reg_fitted %>%
  pull_tflow_fit()

#> parsnip model object
#> 
#> Fit time:  23ms 
#> 
#> Call:  glmnet::glmnet(x = as.matrix(x), y = y, family = "gaussian") 
#> 
#>    Df  %Dev Lambda
#> 1   0  0.00 5.1470
#> 2   2 12.90 4.6900
#> 3   2 24.81 4.2730
#> 4   2 34.69 3.8940
#> 5   2 42.90 3.5480
#> 6   2 49.71 3.2320
#> 7   2 55.37 2.9450
#> 8   2 60.06 2.6840
#> 9   2 63.96 2.4450
#> 10  3 67.26 2.2280
#> 11  3 70.15 2.0300
#> 12  3 72.56 1.8500
#> 13  3 74.55 1.6850
#> 14  3 76.21 1.5360
#> 15  3 77.59 1.3990
#> 16  3 78.73 1.2750
#> 17  3 79.68 1.1620
#> 18  3 80.46 1.0580
#> 19  3 81.12 0.9645
#> 20  3 81.66 0.8788
#> 21  3 82.11 0.8007
#> 22  3 82.49 0.7296
#> 23  4 82.81 0.6648
#> 24  5 83.20 0.6057
#> 25  5 83.60 0.5519
#> 26  6 83.96 0.5029
#> 27  6 84.26 0.4582
#> 28  6 84.51 0.4175
#> 29  6 84.72 0.3804
#> 30  8 84.89 0.3466
#> 31  8 85.14 0.3158
#> 32  8 85.35 0.2878
#> 33  8 85.53 0.2622
#> 34  8 85.68 0.2389
#> 35  8 85.80 0.2177
#> 36  8 85.90 0.1983
#> 37  8 85.98 0.1807
#> 38  9 86.06 0.1647
#> 39  9 86.15 0.1500
#> 40  9 86.22 0.1367
#> 41  9 86.27 0.1246
#> 42  9 86.32 0.1135
#> 43  9 86.36 0.1034
#> 44  9 86.39 0.0942
#> 45  9 86.42 0.0859
#> 46  9 86.44 0.0782
#> 47  9 86.46 0.0713
#> 48  9 86.48 0.0649
#> 49  9 86.49 0.0592
#> 50  9 86.50 0.0539
#> 51  9 86.51 0.0491
#> 52  9 86.52 0.0448
#> 53  9 86.52 0.0408
#> 54 10 86.54 0.0372
#> 55 10 86.60 0.0339
#> 56 10 86.65 0.0309
#> 57 10 86.69 0.0281
#> 58 10 86.73 0.0256
#> 59 10 86.76 0.0233
#> 60 10 86.78 0.0213
#> 61 10 86.80 0.0194
#> 62 10 86.82 0.0177
#> 63 10 86.83 0.0161
#> 64 10 86.84 0.0147
#> 65 10 86.85 0.0134
#> 66 10 86.86 0.0122
#> 67 10 86.87 0.0111
#> 68 10 86.87 0.0101
#> 69 10 86.88 0.0092
#> 70 10 86.88 0.0084
#> 71 10 86.88 0.0076
#> 72 10 86.89 0.0070
#> 73 10 86.89 0.0063
#> 74 10 86.89 0.0058
#> 75 10 86.89 0.0053
#> 76 10 86.89 0.0048
#> 77 10 86.89 0.0044
#> 78 10 86.90 0.0040
#> 79 10 86.90 0.0036