Extract elements of a tidyflow

These functions extract various elements from a tidyflow object. If they do not exist yet, an error is thrown. Most of these steps can only be executed after the tidyflow has been fitted.

pull_tflow_rawdata() returns the complete raw/untrained data.

pull_tflow_split() returns the split object from the function specified in plug_split.

pull_tflow_training() turns the training data from the split. Only works when a split has been specified with plug_split. If prep = TRUE, the preprocessing (either recipe or formula) is applied to the data.

pull_tflow_testing() returns the testing data from the split. Only works when a split has been specified with plug_split If prep = TRUE, the preprocessing (either recipe or formula) is applied to the data.

pull_tflow_preprocessor() returns either the formula or recipe used for preprocessing. Note that in the case of the recipe it returns the function that should be applied to the data to apply the recipe. pull_tflow_prepped_recipe returns the recipe already applied to the training data.

pull_tflow_resample() returns the resample object from the function specified in plug_resample. The resample object does not have the preprocessor applied (either formula or recipe).

pull_tflow_grid() returns the grid data frame from which the tuning parameter was made.

pull_tflow_prepped_recipe() returns the prepped recipe. It is extracted from the mold object returned from pull_tflow_mold().

pull_tflow_mold() returns the preprocessed "mold" object returned from hardhat::mold(). It contains information about the preprocessing, including either the prepped recipe or the formula terms object.

pull_tflow_spec() returns the parsnip model specification.

pull_tflow_fit() returns the parsnip model fit.

pull_tflow_fit_tuning() returns the resample result from model tuning.

pull_tflow_rawdata(x)

pull_tflow_split(x)

pull_tflow_training(x, prep = FALSE)

pull_tflow_testing(x, prep = FALSE)

pull_tflow_resample(x)

pull_tflow_grid(x)

pull_tflow_preprocessor(x)

pull_tflow_prepped_recipe(x)

pull_tflow_mold(x)

pull_tflow_spec(x)

pull_tflow_fit(x)

pull_tflow_fit_tuning(x)

Arguments

x
A tidyflow

prep
A logical stating whether the training/testing data should be returned with the preprocessing step applied (either the formula or the recipe preprocessing). By default it is set to FALSE.

Value

The extracted value from the tidyflow, x, as described in the description section.

Examples

library(parsnip)
library(recipes)
library(rsample)

model <- set_engine(linear_reg(), "lm")

recipe <- ~ recipe(.x, mpg ~ cyl + disp) %>% step_log(disp)

tflow <-
 mtcars %>%
 tidyflow() %>%
 plug_split(initial_split) %>%
 plug_model(model)

recipe_tflow <- plug_recipe(tflow, recipe)
formula_tflow <- plug_formula(tflow, mpg ~ cyl + log(disp))

fit_recipe_tflow <- fit(recipe_tflow)
fit_formula_tflow <- fit(formula_tflow)

pull_tflow_rawdata(fit_recipe_tflow)

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2

pull_tflow_rawdata(fit_formula_tflow)

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2

# The preprocessor is either the recipe function or a formula
pull_tflow_preprocessor(recipe_tflow)

#> <lambda>
#> function (..., .x = ..1, .y = ..2, . = ..1) 
#> recipe(.x, mpg ~ cyl + disp) %>% step_log(disp)
#> <environment: 0x7f8e2de9ad28>
#> attr(,"class")
#> [1] "rlang_lambda_function" "function"             

pull_tflow_preprocessor(formula_tflow)

#> mpg ~ cyl + log(disp)
#> <environment: 0x7f8e2de9ad28>

# The `spec` is the parsnip spec before it has been fit.
# The `fit` is the fit parsnip model.
pull_tflow_spec(fit_formula_tflow)

#> Linear Regression Model Specification (regression)
#> 
#> Computational engine: lm 
#> 

pull_tflow_fit(fit_formula_tflow)

#> parsnip model object
#> 
#> Fit time:  2ms 
#> 
#> Call:
#> stats::lm(formula = ..y ~ ., data = data)
#> 
#> Coefficients:
#> (Intercept)          cyl  `log(disp)`  
#>     70.7902       0.1413      -9.7985  
#> 

# The mold is returned from `hardhat::mold()`, and contains the
# predictors, outcomes, and information about the preprocessing
# for use on new data at `predict()` time.
pull_tflow_mold(fit_recipe_tflow)

#> $predictors
#> # A tibble: 24 x 2
#>      cyl  disp
#>    <dbl> <dbl>
#>  1     6  5.08
#>  2     4  4.68
#>  3     6  5.55
#>  4     8  5.89
#>  5     6  5.42
#>  6     4  4.95
#>  7     6  5.12
#>  8     6  5.12
#>  9     8  5.62
#> 10     8  5.62
#> # … with 14 more rows
#> 
#> $outcomes
#> # A tibble: 24 x 1
#>      mpg
#>    <dbl>
#>  1  21  
#>  2  22.8
#>  3  21.4
#>  4  18.7
#>  5  18.1
#>  6  22.8
#>  7  19.2
#>  8  17.8
#>  9  16.4
#> 10  17.3
#> # … with 14 more rows
#> 
#> $blueprint
#> Recipe blueprint: 
#>  
#> # Predictors: 2 
#>   # Outcomes: 1 
#>    Intercept: FALSE 
#> Novel Levels: FALSE 
#> 
#> $extras
#> $extras$roles
#> NULL
#> 
#> 

# The raw training and testing
pull_tflow_training(fit_recipe_tflow)

#>                     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4          21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Datsun 710         22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive     21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant            18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Merc 230           22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280           19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C          17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE         16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL         17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Cadillac Fleetwood 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Chrysler Imperial  14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Honda Civic        30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla     33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Dodge Challenger   15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin        15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28         13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird   19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9          27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2      26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Ford Pantera L     15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino       19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora      15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E         21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2

pull_tflow_testing(fit_recipe_tflow)

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2

# Or with the preprocessor (recipe/formula) applied
pull_tflow_training(fit_recipe_tflow, prep = TRUE)

#>     mpg cyl     disp
#> 1  21.0   6 5.075174
#> 2  22.8   4 4.682131
#> 3  21.4   6 5.552960
#> 4  18.7   8 5.886104
#> 5  18.1   6 5.416100
#> 6  22.8   4 4.947340
#> 7  19.2   6 5.121580
#> 8  17.8   6 5.121580
#> 9  16.4   8 5.619676
#> 10 17.3   8 5.619676
#> 11 10.4   8 6.156979
#> 12 14.7   8 6.086775
#> 13 30.4   4 4.326778
#> 14 33.9   4 4.264087
#> 15 15.5   8 5.762051
#> 16 15.2   8 5.717028
#> 17 13.3   8 5.857933
#> 18 19.2   8 5.991465
#> 19 27.3   4 4.369448
#> 20 26.0   4 4.789989
#> 21 15.8   8 5.860786
#> 22 19.7   6 4.976734
#> 23 15.0   8 5.707110
#> 24 21.4   4 4.795791

pull_tflow_testing(fit_recipe_tflow, prep = TRUE)

#>    mpg cyl     disp
#> 1 21.0   6 5.075174
#> 2 14.3   8 5.886104
#> 3 24.4   4 4.988390
#> 4 15.2   8 5.619676
#> 5 10.4   8 6.131226
#> 6 32.4   4 4.365643
#> 7 21.5   4 4.788325
#> 8 30.4   4 4.554929

# A useful shortcut is to extract the prepped recipe from the tidyflow
pull_tflow_prepped_recipe(fit_recipe_tflow)

#> Data Recipe
#> 
#> Inputs:
#> 
#>       role #variables
#>    outcome          1
#>  predictor          2
#> 
#> Training data contained 24 data points and no missing data.
#> 
#> Operations:
#> 
#> Log transformation on disp [trained]

# That is identical to
identical(
  pull_tflow_mold(fit_recipe_tflow)$blueprint$recipe,
  pull_tflow_prepped_recipe(fit_recipe_tflow)
)

#> [1] TRUE