Add formula terms to a tidyflow

plug_formula() specifies the terms of the model through the usage of a formula.

drop_formula() removes the formula as well as any downstream objects that might get created after the formula is used for preprocessing, such as terms. Additionally, if the model has already been fit, then the fit is removed.

replace_formula() first removes the formula, then replaces the previous formula with the new one. Any model that has already been fit based on this formula will need to be refit.

plug_formula(x, formula, ..., blueprint = NULL)

drop_formula(x)

replace_formula(x, formula, ..., blueprint = NULL)

Arguments

x
A tidyflow

formula
A formula specifying the terms of the model. It is advised to not do preprocessing in the formula, and instead use a recipe if that is required.

...
Not used.

blueprint
A hardhat blueprint used for fine tuning the preprocessing. If NULL, hardhat::default_formula_blueprint() is used.

Value

The tidyflow x, updated with either a new or removed formula preprocessor.

Details

To fit a tidyflow, one of plug_formula() or plug_recipe() must be specified, but not both.

Examples

# Just for the pipe: %>%
library(tibble)

tflow <-
  mtcars %>%
  tidyflow(seed = 652341) %>% 
  plug_formula(mpg ~ .)

tflow

#> ══ Tidyflow ════════════════════════════════════════════════════════════════════
#> Data: 32 rows x 11 columns
#> Split: None
#> Formula: mpg ~ .
#> Resample: None
#> Grid: None
#> Model: None

drop_formula(tflow)

#> ══ Tidyflow ════════════════════════════════════════════════════════════════════
#> Data: 32 rows x 11 columns
#> Split: None
#> Recipe/Formula: None
#> Resample: None
#> Grid: None
#> Model: None

replace_formula(tflow, mpg ~ disp)

#> ══ Tidyflow ════════════════════════════════════════════════════════════════════
#> Data: 32 rows x 11 columns
#> Split: None
#> Formula: mpg ~ disp
#> Resample: None
#> Grid: None
#> Model: None